modified ImageSlider component
Description: {} This is added at the end of a function which is the expression of the function but it is used without the keyword 'function' and a function name. Inside it has the SliderData component which has the map function attached to it with two parameters, slide and index. An arrow key is typed with an expression of the nameless function returning an img html tag and inside the tag a js code is attached with one of the parameters 'slide' attached to it is an image property which is passed through the map function